Correct measurement equation in Chapter 1.3
Measuring equations in Section 2.1 get the order of the qubits in a position different from the Born's Rule. Even though the result is not affected by such an inversion in any of the examples, it contradicts the formula from Chapter 1.4, Section 4. It is best that Section 2.1 be corrected in order to avoid any sort of confusion for the reader.

This affected lines:
2361 - x and psi switched
2363 - psi switched for x to be coherent with the equation above
2394 - line excluded since the row vector of q_0 is not used anymore, therefore it can be omitted (line left blank for ease of revision)
2395 - q_0 and 0 switched. After the imaginary number, 1 and 0 switched.
2396 to 2398 - q_0 and 0 switched.